+++ src/argo_interp/__init__.py
@@ -6,13 +6,17 @@ from .cycle import InterpolationModel as InterpolationModel
 from .cycle import ModelKwargs as ModelKwargs
 from .cycle import ModelSettings as ModelSettings
 from .cycle import SensorAccuracy as SensorAccuracy
+from .field import FieldProfile as FieldProfile
+from .field import LocalWeightedField as LocalWeightedField
 
 __all__ = [
     "CycleBatch",
     "CycleCollection",
     "CycleIndex",
+    "FieldProfile",
     "InterpolationAdapters",
     "InterpolationModel",
+    "LocalWeightedField",
     "ModelKwargs",
     "ModelSettings",
     "SensorAccuracy",